requirements.txt
asgiref==3.6.0
Django==4.1.7
pip==23.0.1
setuptools==65.5.1
sqlparse==0.4.3
tzdata==2022.7
wheel==0.38.4
djangorestframework==3.14.0
python-dotenv==0.21.1
psycopg2==2.9.5